kafka-pipeline/requirements.txt
# ===== FILE 1: kafka-pipeline/requirements.txt =====
kafka-python==2.0.2
psycopg2-binary==2.9.9
requests==2.31.0